Move cem stuff into cem folder.
