+++ cem/cem.c
