+++ cem/consts.h
