+++ cem/globals.h
